Refine admin report trend helpers
